chore(jans-cedarling): extract `Request` structure to authz module
